refactor(components): 共有スキームに合わせ web の app UI を責務カテゴリへ再編
packages/components の責務スキームを web の app 固有 UI にも適用し、
components/ui/{責務カテゴリ}/ に集約。

- content-pagination→navigation、empty-state→feedback、
  inline-tag-filter/search-input→inputs、language-switcher/theme-toggle→actions、
  mobile-filter-sheet→overlays、tag-pill→display
- content/ errors/ seo/ は web 専用概念として据え置き
- deep import ~23 箇所を新 path へ更新

typecheck / lint / lint:boundaries / build pass。
