+++ apps/web/src/components/ui/inputs/inline-tag-filter.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { TagPill } from '@/components/ui/tag-pill';
+import { TagPill } from '@/components/ui/display/tag-pill';
 import { Button } from '@dayopt/components';
 import { X } from 'lucide-react';
 import { useState } from 'react';
