+++ apps/web/src/features/blog/components/FilteredBlogClient.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { ContentHeader } from '@/components/content/ContentHeader';
-import { ContentPagination } from '@/components/ui/content-pagination';
-import { EmptyState } from '@/components/ui/empty-state';
-import { SearchInput } from '@/components/ui/search-input';
+import { EmptyState } from '@/components/ui/feedback/empty-state';
+import { SearchInput } from '@/components/ui/inputs/search-input';
+import { ContentPagination } from '@/components/ui/navigation/content-pagination';
 import { Rss, Search } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 import { useSearchParams } from 'next/navigation';
