+++ apps/web/src/features/blog/components/PostCard.tsx
@@ -1,4 +1,4 @@
-import { TagPill } from '@/components/ui/tag-pill';
+import { TagPill } from '@/components/ui/display/tag-pill';
 import { Link } from '@/platform/i18n/navigation';
 import { Heading } from '@dayopt/components';
 import { BlogPostMeta } from '../lib/blog';
